+++ src/public_data/operation/get_company_profile.rs
@@ -2,9 +2,8 @@ use reqwest::StatusCode;
 use thiserror::Error;
 use typed_builder::TypedBuilder;
 
-use crate::{types, unrecognised_response::UnrecognisedResponse};
-
 use super::{CompaniesHousePublicDataOperation, CompaniesHousePublicDataOperationError};
+use crate::{types::CompanyProfile, unexpected_status::UnexpectedStatusError};
 
 /// Get the basic company information
 #[derive(TypedBuilder)]
@@ -14,53 +13,37 @@ pub struct GetCompanyProfile {
 }
 
 #[derive(Debug, Error)]
-pub enum GetCompanyProfileError {
+pub enum GetCompanyProfileStatusError {
     #[error("Unauthorized")]
     Unauthorized,
     #[error("Not found")]
     NotFound,
-    #[error("Operation error")]
-    OperationError(#[from] CompaniesHousePublicDataOperationError),
-    #[error("Bad JSON")]
-    SerdeJson(#[from] reqwest::Error),
-    #[error("UnrecognisedResponse {0:?}")]
-    UnrecognisedResponse(#[from] UnrecognisedResponse),
+    #[error(transparent)]
+    UnexpectedStatus(#[from] UnexpectedStatusError),
 }
 
 impl CompaniesHousePublicDataOperation for GetCompanyProfile {
-    type Error = GetCompanyProfileError;
-    type Data = types::CompanyProfile;
+    type StatusError = GetCompanyProfileStatusError;
+    type Data = CompanyProfile;
 
     fn build_request(
         &self,
         base_url: &str,
         client: &reqwest::Client,
-    ) -> Result<reqwest::Request, CompaniesHousePublicDataOperationError> {
+    ) -> Result<reqwest::Request, CompaniesHousePublicDataOperationError<Self::StatusError>> {
         Ok(client
             .get(format!("{base_url}/company/{}", self.company_number))
             .build()?)
     }
 
-    async fn handle_response(
-        &self,
-        response: reqwest::Response,
-    ) -> Result<Self::Data, Self::Error> {
-        match response.status() {
-            StatusCode::UNAUTHORIZED => return Err(Self::Error::Unauthorized),
-            StatusCode::NOT_FOUND => return Err(Self::Error::NotFound),
-            StatusCode::OK => {}
-            _ => {
-                return Err(Self::Error::UnrecognisedResponse(
-                    UnrecognisedResponse::from_response(response).await,
-                ))
-            }
-        };
-
-        let value: Self::Data = response
-            .json()
-            .await
-            .map_err(CompaniesHousePublicDataOperationError::Reqwest)?;
-
-        Ok(value)
+    fn handle_status(&self, status_code: StatusCode) -> Result<(), Self::StatusError> {
+        match status_code {
+            StatusCode::OK => Ok(()),
+            StatusCode::UNAUTHORIZED => Err(Self::StatusError::Unauthorized),
+            StatusCode::NOT_FOUND => Err(Self::StatusError::NotFound),
+            status_code => Err(Self::StatusError::UnexpectedStatus(UnexpectedStatusError {
+                status_code,
+            })),
+        }
     }
 }